requirement.txt
alipay-sdk-python==3.3.202
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
django-haystack==2.8.1
django-redis==4.10.0
django-tinymce==2.8.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
jieba==0.39
kombu==4.6.5
more-itertools==7.2.0
mutagen==1.43.0
mysqlclient==1.4.4
Pillow==6.2.1
py3Fdfs==2.2.0
pyasn1==0.4.8
pycrypto==2.6.1
PyMySQL==0.9.3
pytz==2019.3
redis==3.3.11
requests==2.22.0
rsa==4.0
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
Whoosh==2.7.4
zipp==0.6.0
